chore: setup github client before running tasks
