+++ lib/release.js
@@ -6,10 +6,10 @@ const semVer = require('semver')
 
 class Release {
 
-  constructor({pre = null, client = null} = {}) {
+  constructor({pre = null, github = null} = {}) {
     this.body = null
     this.changes = null
-    this.client = client
+    this.client = github
     this.commits = null
     this.credits = null
     this.data = null
