+++ lib/tasks/start.js
@@ -1,7 +1,6 @@
 'use strict'
 
 // Ours
-const connect = require('../connect')
 const {create: newRelease} = require('../release')
 const {getRepo} = require('../repo')
 const inSync = require('./in-sync')
@@ -11,8 +10,6 @@ module.exports = config => async (release = newRelease(config)) => {
     getRepo(),
     inSync(config)(release)
   ])
-
-  release.client = await connect(config.progress)
 
   return Object.assign(release, details)
 }
